+++ packages/mobx-state-tree/__tests__/core/identifier.test.ts
@@ -1,4 +1,14 @@
-import { types, tryResolve, resolvePath, cast, SnapshotOrInstance } from "../../src"
+import {
+    types,
+    tryResolve,
+    resolvePath,
+    cast,
+    SnapshotOrInstance,
+    IAnyModelType,
+    IAnyComplexType,
+    resolveIdentifier,
+    getIdentifier
+} from "../../src"
 
 if (process.env.NODE_ENV !== "production") {
     test("#275 - Identifiers should check refinement", () => {
@@ -56,9 +66,9 @@ test("#158 - #88 - Identifiers should accept any string character", () => {
 })
 test("#187 - identifiers should not break late types", () => {
     expect(() => {
-        const MstNode: any = types.model("MstNode", {
+        const MstNode = types.model("MstNode", {
             value: types.number,
-            next: types.maybe(types.late(() => MstNode))
+            next: types.maybe(types.late((): IAnyModelType => MstNode))
         })
     }).not.toThrow()
 })
@@ -144,10 +154,10 @@ if (process.env.NODE_ENV !== "production") {
 }
 
 test("it can resolve through refrences", () => {
-    const Folder: any = types.model("Folder", {
+    const Folder = types.model("Folder", {
         type: types.literal("folder"),
         name: types.identifier,
-        children: types.array(types.late(() => types.union(Folder, SymLink)))
+        children: types.array(types.late((): IAnyComplexType => types.union(Folder, SymLink)))
     })
     const SymLink = types.model({
         type: types.literal("link"),
@@ -199,4 +209,56 @@ test("it can resolve through refrences", () => {
     expect(() => resolvePath(root, "/children/3/target/children/0").name).toThrow(
         "[mobx-state-tree] Failed to resolve reference 'e' to type 'Folder' (from node: /children/3/target)"
     )
+})
+
+test("#1019", () => {
+    let calls = 0
+    function randomUuid() {
+        calls++
+        let pattern = "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx"
+        return pattern.replace(/[xy]/g, function(cc) {
+            const r = (Math.random() * 16) | 0,
+                v = cc === "x" ? r : (r & 0x3) | 0x8
+            return v.toString(16)
+        })
+    }
+
+    const TOptionalId = types.optional(
+        types.refinement(types.identifier, identifier =>
+            /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/.test(identifier)
+        ),
+        randomUuid
+    )
+
+    const CommentModel = types.model("CommentModel", {
+        uid: TOptionalId
+    })
+
+    const CommentStore = types
+        .model("CommentStore", {
+            items: types.array(CommentModel)
+        })
+        .actions(self => ({
+            test1() {
+                expect(calls).toBe(0)
+                const comment = CommentModel.create({})
+                expect(calls).toBe(1)
+
+                self.items.push(comment)
+                const item = resolveIdentifier(CommentModel, self.items, comment.uid)
+                const item2 = self.items.find(i => i.uid === comment.uid)
+                expect(item).toBe(item2)
+            }
+        }))
+
+    const c = CommentStore.create({})
+    c.test1()
+})
+
+test("#1019-2", () => {
+    const Item = types.model({
+        id: types.optional(types.identifier, "dummykey")
+    })
+    expect(getIdentifier(Item.create())).toBe("dummykey")
+    expect(Item.create().id).toBe("dummykey")
 })