+++ packages/mobx-state-tree/src/core/type/type.ts
@@ -157,7 +157,7 @@ export type TAndInterface<T, I> = (Exclude<T, Primitives> & I) | Extract<T, Prim
 
 export interface IComplexType<C, S, T> extends IType<C, S, T> {
     // fake, only used for typing
-    readonly $complexType: undefined
+    readonly "!!complexType": undefined
 
     create(
         snapshot?: C,
@@ -218,7 +218,7 @@ export type SnapshotOrInstance<T> = SnapshotIn<T> | Instance<T>
  * @private
  */
 export abstract class ComplexType<C, S, T> implements IComplexType<C, S, T> {
-    readonly $complexType = undefined
+    readonly "!!complexType" = undefined
 
     readonly isType = true
     readonly name: string