+++ packages/mobx-state-tree/src/types/utility-types/optional.ts
@@ -25,7 +25,7 @@ export type IFunctionReturn<T> = () => T
  * @internal
  * @private
  */
-export type IOptionalValue<C, S, T> = C | S | T | IFunctionReturn<C | S | T>
+export type IOptionalValue<C, S, T> = C | S | IFunctionReturn<C | S | T>
 
 /**
  * @internal
@@ -55,8 +55,8 @@ export class OptionalValue<C, S, T> extends Type<C, S, T> {
 
     instantiate(parent: INode, subpath: string, environment: any, value: S): INode {
         if (typeof value === "undefined") {
-            const defaultSnapshot = this.getDefaultValueSnapshot()
-            return this.type.instantiate(parent, subpath, environment, defaultSnapshot)
+            const defaultInstanceOrSnapshot = this.getDefaultInstanceOrSnapshot()
+            return this.type.instantiate(parent, subpath, environment, defaultInstanceOrSnapshot)
         }
         return this.type.instantiate(parent, subpath, environment, value)
     }
@@ -64,22 +64,30 @@ export class OptionalValue<C, S, T> extends Type<C, S, T> {
     reconcile(current: INode, newValue: any): INode {
         return this.type.reconcile(
             current,
-            this.type.is(newValue) && newValue !== undefined ? newValue : this.getDefaultValue()
+            this.type.is(newValue) && newValue !== undefined
+                ? newValue
+                : this.getDefaultInstanceOrSnapshot()
         )
     }
 
-    private getDefaultValue() {
-        const defaultValue =
+    getDefaultInstanceOrSnapshot() {
+        const defaultInstanceOrSnapshot =
             typeof this.defaultValue === "function" ? this.defaultValue() : this.defaultValue
-        if (typeof this.defaultValue === "function") typecheckInternal(this, defaultValue)
-        return defaultValue
+
+        // while static values are already snapshots and checked on types.optional
+        // generator functions must always be rechecked just in case
+        if (typeof this.defaultValue === "function") {
+            typecheckInternal(this, defaultInstanceOrSnapshot)
+        }
+
+        return defaultInstanceOrSnapshot
     }
 
     public getDefaultValueSnapshot() {
-        const defaultValue = this.getDefaultValue()
-        return isStateTreeNode(defaultValue)
-            ? getStateTreeNode(defaultValue).snapshot
-            : defaultValue
+        const instanceOrSnapshot = this.getDefaultInstanceOrSnapshot()
+        return isStateTreeNode(instanceOrSnapshot)
+            ? getStateTreeNode(instanceOrSnapshot).snapshot
+            : instanceOrSnapshot
     }
 
     isValidSnapshot(value: any, context: IContext): IValidationResult {
@@ -98,21 +106,13 @@ export class OptionalValue<C, S, T> extends Type<C, S, T> {
 
 export function optional<C, S, T>(
     type: IComplexType<C, S, T>,
-    defaultValueOrFunction: C | S | T
+    defaultValueOrFunction: C | S | (() => C | S | T)
 ): IComplexType<C | undefined, S, T> & OptionalProperty
 export function optional<C, S, T>(
-    type: IComplexType<C, S, T>,
-    defaultValueOrFunction: () => C | S | T
-): IComplexType<C | undefined, S, T> & OptionalProperty
+    type: IType<C, S, T>,
+    defaultValueOrFunction: C | S | (() => C | S | T)
+): IType<C | undefined, S, T> & OptionalProperty
 
-export function optional<C, S, T>(
-    type: IType<C, S, T>,
-    defaultValueOrFunction: C | S | T
-): IType<C | undefined, S, T> & OptionalProperty
-export function optional<C, S, T>(
-    type: IType<C, S, T>,
-    defaultValueOrFunction: () => C | S | T
-): IType<C | undefined, S, T> & OptionalProperty
 /**
  * `types.optional` can be used to create a property with a default value.
  * If the given value is not provided in the snapshot, it will default to the provided `defaultValue`.
@@ -132,19 +132,30 @@ export function optional<C, S, T>(
  * @export
  * @alias types.optional
  */
-export function optional(type: IAnyType, defaultValueOrFunction: any): IAnyType & OptionalProperty {
+export function optional<C, S, T>(
+    type: IType<C, S, T>,
+    defaultValueOrFunction: C | S | (() => C | S | T)
+): IType<C | undefined, S, T> & OptionalProperty {
+    // make sure we never pass direct instances
+    if (typeof defaultValueOrFunction !== "function" && isStateTreeNode(defaultValueOrFunction)) {
+        fail(
+            "default value cannot be an instance, pass a snapshot or a function that creates an instance/snapshot instead"
+        )
+    }
+
     if (process.env.NODE_ENV !== "production") {
         if (!isType(type))
             fail("expected a mobx-state-tree type as first argument, got " + type + " instead")
-        const defaultValue =
-            typeof defaultValueOrFunction === "function"
-                ? defaultValueOrFunction()
-                : defaultValueOrFunction
-        const defaultSnapshot = isStateTreeNode(defaultValue)
-            ? getStateTreeNode(defaultValue).snapshot
-            : defaultValue
-        typecheckInternal(type, defaultSnapshot)
+
+        // we only check default values if they are passed directly
+        // if they are generator functions they will be checked once they are generated
+        // we don't check generator function results here to avoid generating a node just for type-checking purposes
+        // which might generate side-effects
+        if (typeof defaultValueOrFunction !== "function") {
+            typecheckInternal(type, defaultValueOrFunction)
+        }
     }
+
     const ret = new OptionalValue(type, defaultValueOrFunction)
     return ret as typeof ret & OptionalProperty
 }