+++ packages/mst-middlewares/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    displayName: "test",
+    testEnvironment: "node",
+    transform: {
+        "^.+\\.tsx?$": "ts-jest",
+        "^.+\\.jsx?$": "babel-jest"
+    },
+    testRegex: ".*\\.test\\.tsx?$",
+    moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json"]
+}